support SNI in output plugin
Certain deployments that use a proxy may need to use SNI.  This
patch adds a new parameter `sni_hostname` to the `<server>` section
for the output plugin.  This can only be used if the version of
openssl used supports it.  A warning will be issued if openssl does
not support SNI and `sni_hostname` is attempted to be used.